wip: structuring
[skip ci]
