+++ src/typegraph/rust/src/error.rs
@@ -1,0 +1,3 @@
+pub use crate::wasm::core::Error;
+
+pub type Result<T> = std::result::Result<T, Error>;
