+++ src/typegraph/rust/src/lib.rs
@@ -1,2 +1,9 @@
-#[allow(unused)]
-mod wasm;
+mod error;
+mod host;
+
+pub mod runtimes;
+pub mod t;
+
+pub mod wasm; // FIXME
+
+pub use error::{Error, Result};
